Move JuttleJob to its own file.
This is in anticipation of making it a base class from which you can
have websocket and immediate based jobs.
